pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Automate data captures at the freestate oil weighbridge"
authors = ["Raku Dreemurr <rakudreemurr@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
pytesseract = "^0.3.10"
pdf2image = "^1.16.3"
Kivy = "^2.2.1"
portray = "^1.8.0"
mysql-connector = "^2.2.9"

[tool.poetry.scripts]
portray = "portray.server.serve"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
